Organized
I organized some files and altered the wave animation a bit to make it flow better.
