+++ DynamicAnimations/src/components/background/background.css
@@ -1,0 +1,3 @@
+.Background svg {
+  shape-rendering: crispEdges;
+}
